# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The jinshuzhai bookstore api server.",
        "title": "High-Performance-Online-Bookstore",
        "termsOfService": "https://github.com/zouyonghe",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/zouyonghe",
            "email": "1259085392z@gmail.com"
        },
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.7.0"
    },
    "host": "127.0.0.1:8081",
    "basePath": "/v1",
    "paths": {
...
        "/user/admin": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List users account by specified username format include id, username, role, etc. The password hash is never returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user/admin"
                ],
                "summary": "List users account by specified username format.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "fuzzy username",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number, default 1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, default 10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{\"totalCount\":1,\"userList\":[{\"id\":1,\"username\":\"admin\",\"ShortId\":\"5P9Ia4QnR\",\"role\":\"admin\",\"createdAt\":\"2021-04-18 15:40:33\",\"updatedAt\":\"2021-04-18 15:40:33\"}]}}",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerListResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.UserInfo": {
            "type": "object",
            "properties": {
                "ShortId": {
                    "type": "string"
                },
                "UserID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.SwaggerListResponse": {
            "type": "object",
            "properties": {
                "totalCount": {
                    "type": "integer"
                },
                "userList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserInfo"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}